README.md
This script accesses the webcam and converts the captured frame from BGR
to HSV and the HSV threshold is adjusted such that only blue ,red and
green objects are shown on the screen.(INDIVIDUALLY)

This script can also be used to detect other colored objects. All that
has to be done is to just change the HSV threshold ie change the
hue-saturation-value according to the color one wants to track.